bug fix
IE XMLSerializer produces some oddities when non-standard name-spaces
are used which throws an error when trying to load the image. No way to
catch it, so we now do the cleaning of namespace declarations for
everyone. This may reduce the length of our dataString anyway, so maybe
not a bad thing.